+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.0.6](https://github.com/torlenor/redseligg/releases/tag/0.0.6) (TBD)
+
+**Implemented enhancements:**
+
+**New storage support:**
+
+**New platforms:**
+
+**New plugins:**
+
+- Archive plugin: Stores all messages with their timestamps in the storage.
+
 ## [0.0.5](https://github.com/torlenor/redseligg/releases/tag/0.0.5) (2020-05-22)
 
 **Implemented enhancements:**